# Optimize ptrSetter