Before:

```
alex@penguin ~/p/go-fixedwidth> go test -bench='BenchmarkUnmarshal_MixedData_1000'
goos: linux
goarch: amd64
pkg: github.com/ianlopshire/go-fixedwidth
BenchmarkUnmarshal_MixedData_1000-4     	    5000	    390835 ns/op
BenchmarkUnmarshal_MixedData_100000-4   	      50	  33693614 ns/op
PASS
ok  	github.com/ianlopshire/go-fixedwidth	3.740s
alex@penguin ~/p/go-fixedwidth> go test -bench='BenchmarkUnmarshal_MixedData_1000'
goos: linux
goarch: amd64
pkg: github.com/ianlopshire/go-fixedwidth
BenchmarkUnmarshal_MixedData_1000-4     	    5000	    399255 ns/op
BenchmarkUnmarshal_MixedData_100000-4   	      50	  35145768 ns/op
PASS
ok  	github.com/ianlopshire/go-fixedwidth	3.857s
```

After:

```
alex@penguin ~/p/go-fixedwidth> go test -bench='BenchmarkUnmarshal_MixedData_1000'
goos: linux
goarch: amd64
pkg: github.com/ianlopshire/go-fixedwidth
BenchmarkUnmarshal_MixedData_1000-4     	    5000	    327261 ns/op
BenchmarkUnmarshal_MixedData_100000-4   	      50	  30777707 ns/op
PASS
ok  	github.com/ianlopshire/go-fixedwidth	3.272s
alex@penguin ~/p/go-fixedwidth> go test -bench='BenchmarkUnmarshal_MixedData_1000'
goos: linux
goarch: amd64
pkg: github.com/ianlopshire/go-fixedwidth
BenchmarkUnmarshal_MixedData_1000-4     	    5000	    348843 ns/op
BenchmarkUnmarshal_MixedData_100000-4   	      50	  31734454 ns/op
PASS
ok  	github.com/ianlopshire/go-fixedwidth	3.431s
```